==> 2. Aggregate bat data to night with weather.Rmd <==


title: "Aggregate bat acoustic data to night with weather"
output: html_document
date: "2022-11-25"
---

# Input :  The cleaned manual acoustic analysis data from the Marker 2020 study 
MarkerManual_cleaned_manualid_guild_sitenames.csv created 14.05.2023 

# Outputs: 
            1. All bat activity aggregated to night per site 
           AND all bat activity aggregated to night per site - BINARY (1/0) 
           
           2. All guild and behavior bat activity aggregated to night per site
           AND all bat activity aggregated to night per site - BINARY (1/0) 
           
*Data sets for downstream analysis* : 
 - guild_behavior_summarytable_site.csv
 - totalbatpass_summarytable_withbinary_batpass_night_aggregated_site data.csv
 - nightlyaggregated_zeros_binary_behavior and guild.csv
 - nightlyaggregatedMarker2020_totalbats_zeroinserted_weather_binary.csv
 - guild_behavior_batpass_summarytable_trimmed MRE social and met tower.csv
 
           
## Set up work environment 
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = F)
knitr::opts_chunk$set(tidy = "styler")
options(knitr.table.format = "html")


#Load libraries 
library(knitr)
library(data.table)
library(tidyverse)
library(beepr)
library(lubridate)
library(purrr)
library(renv)
library(stringr)
library(janitor)
library(anytime)
library(kableExtra)
library(papeR)
library(skimr)
library(vtable)


#### Import data  ####

input <- "C:/Users/apmc/OneDrive - Norwegian University of Life Sciences/2. Marker 2019-2020/Marker 2022/SecondDraftAnalyses/Input"

dataset1 <- "MarkerManual_cleaned_manualid_guild_sitenames.csv"
dataset2 <- "Marker_night_inventory_0704.2022_edit.csv"
dataset3 <- "nightlyaggregate_wind_temp.csv"

path1 <- str_c(input, "/", dataset1)
path2 <- str_c(input, "/", dataset2)
path3 <- str_c(input, "/", dataset3)

bats1 <- read_csv(path1) # 19438 bat passes with 22 vars
active.nights <- read.csv(path2, sep = ";") # # 951 obs of 3 variables
weather.nights <- read_csv(path3) # 93 obs of 8 vars


# for Reed
output <- "C:/Users/apmc/OneDrive - Norwegian University of Life Sciences/2. Marker 2019-2020/Marker 2022/SecondDraftAnalyses/Reed/Outputs"

## 
 file.name <- "Step2.Aggregating bat passes to night"
# 
 todays_date <- Sys.Date()
# 
dir.name <- str_c(output,"/", file.name, "_", todays_date)
 dir.name
# 
 dir.create(dir.name) # be careful not to run over an already created directory 

output_today <- dir.name

setwd(output_today)

getwd()

```


## Summary tables of the unaggregated dataset 
```{r}
# make factor levels arranged in orders that are more meaningful and insert a taxa column with more specific manual id names. 

names(bats1)
cols <- c("manual.id" ,  "behavior"  ,  "guild"     ,  
          "Site"     ,   "Habitat"   ,  "Locality"  ,  "Facility")
bats1[cols] <- lapply(bats1[cols], factor)
bats1$taxa <- bats1$manual.id
levels(bats1$taxa)
summary(bats1$taxa)

# "BABA" "EPNI" "LR1"  "LR2"  "MR1"  "NYNO" "NoID" "PAUR" "PINA" "PIPY" "SR1"  "VEMU"
levels(bats1$taxa) <- list("Barbastella barbastellus" = "BABA",
                           "Plecotus auritus" = "PAUR",
                           "Short range group 1" = "SR1",
                           "Eptesicus nilssonii" = "EPNI",
                           "Nyctalus noctula" = "NYNO",
                           "Vespertilio murinus" = "VEMU",
                           "Long range group 1" = "LR1",
                           "Long range group 2" = "LR2",
                           "Medium range group1" = "MR1",
                           "Pipistrellus nathusii" = "PINA",
                           "Pipistrellus pygmaeus" = "PIPY",
                            "Unknown bat" = "NoID") 
summary(bats1$taxa)
# Barbastella barbastellus         Plecotus auritus      Short range group 1      Eptesicus nilssonii         Nyctalus noctula 
#                       10                      553                     3578                    14015                      182 
#      Vespertilio murinus       Long range group 1       Long range group 2      Medium range group1    Pipistrellus nathusii 
#                      168                      132                      221                       22                        9 
#    Pipistrellus pygmaeus              Unknown bat 
#                       64                      484 

levels(bats1$guild) <- list("SRE" = "SRE", "MRE" = "MRE", "LRE" = "LRE", "NoID" = "NoID")
summary(bats1$guild)

levels(bats1$Locality)
levels(bats1$Locality) <-list("Turbine2" = "Turbine2" , "Turbine4" = "Turbine4",  "Turbine8" = "Turbine8", "MeteorologicalTower" = "MeteorologicalTower", "Turbine9" = "Turbine9", "Turbine10" = "Turbine10", "Turbine11" = "Turbine11", "Turbine14" = "Turbine14") 
summary(bats1$Locality)

# Printable summary tables 
kbl(summarize(bats1, type = "factor", variables = "manual.id"))  %>%  kable_styling()
kbl(summarize(bats1, type = "factor", variables = "taxa"))  %>%  kable_styling()
kbl(summarize(bats1, type = "factor", variables = "guild"))  %>%  kable_styling()
kbl(summarize(bats1, type = "factor", variables = "Site"))  %>%  kable_styling()
kbl(summarize(bats1, type = "factor", variables = "Habitat"))  %>%  kable_styling()
kbl(summarize(bats1, type = "factor", variables = "Facility"))  %>%  kable_styling()

names(bats1)

```

##
Create a simple version of this dataset for AFVV
```{r}
bats1$AUTO.ID. <- factor(bats1$AUTO.ID.)
levels(bats1$AUTO.ID.)

# "BARBAR" "EPTNIL" "MYOBRA" "MYODAU" "MYOMYS" "MYONAT" "NYCNOC"
# "NoID"   "PIPNAT" "PIPPIP" "PIPPYG" "PLEAUR" "VESMUR"

b1 <- bats1 %>% select(IN.FILE, 
                       date, TIME, HOUR, 
                       night, TIME.12, HOUR.12,
                       AUTO.ID.,
                       manual.id,
                       taxa,
                       PULSES, 
                       MATCH.RATIO, 
                       guild, 
                       Site, 
                       Habitat,
                       ) %>% 
  rename(auto.id = AUTO.ID.) %>% 
  mutate(auto.guild = factor(case_when(
    auto.id %in% c("EPTNIL", "NYCNOC", "VESMUR") ~ "LRE", 
    auto.id %in% c("PIPNAT", "PIPPIP", "PIPPYG") ~ "MRE", 
    auto.id %in% c("BARBAR", "MYOBRA", "MYODAU", 
                   "MYOMYS", "MYONAT", "PLEAUR") ~ "SRE", 
    auto.id == "NoID" ~ "NoID"
    ))) 
summary(b1)
levels(b1$manual.id)

output_today
# "C:/Users/apmc/OneDrive - Norwegian University of Life Sciences/2. Marker 2019-2020/Marker 2022/SecondDraftAnalyses/Reed/Outputs/Step2.Aggregating bat passes to night_2024-03-26"

write.csv(b1, file = file.path(output_today, "Marker_Bats_Manual_Auto_IDS.csv"))

kbl(names(b1)) %>% kable_styling()
test <- b1 %>% select(IN.FILE) %>% distinct()
```

bats1 <- bats1 %>% mutate(auto_guild = case_when(
  auto %in% c("EPNI", "NYNO", "VEMU", "LR1", "LR2") ~ "LRE",
  auto %in% c("PIPY", "PINA", "MR1") ~ "MRE",
  auto %in% c("BABA", "SR1", "PAUR") ~ "SRE",
  auto == "NoID" ~ "NoID"

## Aggregate bat passes to night per site
- Insert zeros
- merge with night aggregated weather data 
```{r}
# Format the date columns 
bats1 <- bats1 %>% mutate(date = as.Date(DATE),
                           night = as.Date(DATE.12))

bat_sitemap <- bats1 %>% select(Site, Habitat, Facility, Locality) %>% distinct() 

bats_nights_sum <- bats1 %>% group_by(Site, night) %>% dplyr::summarize(batpass = sum(n())) 
bats_nights_sum1 <- left_join(bats_nights_sum, bat_sitemap)
summary(bats_nights_sum1)
summary(bats_nights_sum1$Site)

# write.csv(bats_nights_sum1, "bat pass aggregated by night wout zeronights.csv")


# Insert zero activity nights
head(active.nights)

active.nights$Site <- gsub("C", "N", active.nights$Site)
active.nights$Site <- gsub("MetA", "Met45", active.nights$Site)
active.nights$Site <- gsub("MetB", "Met95", active.nights$Site)
active.nights$Site <- as.factor(active.nights$Site)
active.nights$night <- as.IDate(active.nights$night, '%d.%m.%Y')
active.nights$night <- as.Date(active.nights$night)
summary(active.nights)

dat <- bats_nights_sum1
summary(dat)

summary(dat$Site) # Number of nights with recorded bat activity 
# Met45 Met95   N02   N04   N08   N09   N10   N11   N14   P02   P04   P08   P09   P10   P11   P14 
#    55    17    54    41    49    46    50    55    51    47    52    49    55    49    54    54 

#---- ADDING MISSING DAYS 
# Adjust dataset to include nights when detectors were active but no data was collected (zero nights)

head(active.nights)
summary(active.nights$Site)

nights <- unique(active.nights$night) # 91 unique nights of survey 

temp<-expand.grid(Site = unique(dat$Site),  night = nights) #----expanded data
temp$night <- as.character(temp$night)
temp$night <-  as.Date(temp$night)

active.nights$Site <- as.factor(active.nights$Site)
active.nights$night <-  as.Date(active.nights$night)

str(temp)
str(active.nights)

temp2<-merge(temp, active.nights) # 951 obs of 3 variables 

summary(temp2$Site) # number of monitoring nights per site 
# Met45 Met95   N02   N04   N08   N09   N10   N11   N14   P02   P04   P08   P09   P10   P11   P14 
#    70    70    59    44    60    48    60    60    60    60    60    60    60    60    60    60 

temp3<-merge(temp2,dat,all.x=TRUE) # 851 obs of 7 vars 
summary(temp3$Site)# number of nights per guild per site plot 
# Met45 Met95   N02   N04   N08   N09   N10   N11   N14   P02   P04   P08   P09   P10   P11   P14 
#    70    70    59    44    60    48    60    60    60    60    60    60    60    60    60    60 

summary(temp3) # 173 zero nights (NAs) introduced 

# reformat factor columns and then replace NA bat passes with zeros 
nbats1 <- temp3 

# batpasses converted to zeros but NAs pop up in the factor columns
nbats2 <- nbats1 %>% select(-c(active.night, Habitat, Facility, Locality)) 
summary(nbats2)  

#Replace NAs with zeros 
nbats2[is.na(nbats2)] <- 0 
summary(nbats2)

nbats3 <- left_join(nbats2, bat_sitemap) 
summary(nbats3)

# write.csv(nbats3, "nightlyaggregatedMarker2020_totalbats_zeroinserted.csv") 

```

## Aggregate the hourly weather datasets to nightly 
```{r}

head(weather.nights)

nbats4 <- left_join(nbats3, weather.nights, by = "night")

summary(nbats4) # There is an index column that can be removed 
nbats4 <- nbats4 %>% select(-"...1")

```


## Create binary bat pass value (nights with and without activity)
```{r}

bats2 <- nbats4 %>% rename(Batpass = batpass)
bats2$batpass01 <- bats2$Batpass
names(bats2)
 # [1] "Site"        "night"       "batpass_sum" "Habitat"     "Facility"    "avg_wind"    "min_wind"    "max_wind"   
 # [9] "avg_temp"    "min_temp"    "max_temp"    "batpass01" 

bats2$batpass01[bats2$batpass01 >1] <- 1
summary(bats2)
head(bats2)
tail(bats2)

# * Main results
# write.csv(bats2, "nightlyaggregatedMarker2020_totalbats_zeroinserted_weather_binary.csv")


# last rounds of housekeeping
bats2$wind <- bats2$avg_wind
bats2$temp <- bats2$avg_temp
bats2$jnight <- yday(bats2$night)
# Aggregation table 
# I am keeping the night (as.Date) variable for now because it is easier to recreate jnight from night than the other way around. 

agg_tbl2 <- bats2 %>% group_by(night, Site) %>% 
  dplyr::summarise(across(c(Batpass, batpass01, wind, temp), list(sum = sum, mean = mean, min = min, max=max, length = length)))

dim(agg_tbl2)
# 951  22
head(agg_tbl2)
tail(agg_tbl2)

#Remove variables that will not be needed
names(agg_tbl2)

#  [1] "jnight"           "Site"             "Batpass_sum"      "Batpass_mean"     "Batpass_min"      "Batpass_max"     
#  [7] "Batpass_length"   "batpass01_sum"    "batpass01_mean"   "batpass01_min"    "batpass01_max"    "batpass01_length"
# [13] "wind_sum"         "wind_mean"        "wind_min"         "wind_max"         "wind_length"      "temp_sum"        
# [19] "temp_mean"        "temp_min"         "temp_max"         "temp_length"     


df2C <- agg_tbl2 %>% select(-c(Batpass_min, Batpass_length, batpass01_mean, batpass01_min, batpass01_max,
                                wind_sum, wind_length, temp_sum , temp_length ))

dim(df2C) 
# 951  13
head(df2C)
tail(df2C)


#Make a proportion variable
df2C$batpass_prop <- df2C$batpass01_sum/df2C$batpass01_length

head(df2C)
tail(df2C)
dim(df2C)
summary(df2C) 
sum(df2C$Batpass_sum)
# 19438 we have the correct number of bat passes so this looks good! 


# write.csv(df2C,  "totalbatpass_summarytable_withbinary_batpass_night_aggregated.csv")

# Add back the other site data 
df3C <- left_join(df2C, bat_sitemap)
summary(df3C)
# 951 obs of 17 vars 

# * Main results 
# write.csv(df3C, "totalbatpass_summarytable_withbinary_batpass_night_aggregated_site data.csv")

```


## Quality control plots 
```{r}
names(df3C)

#batpass_prop	versus	jnight
# save plot to file 
bmp(file.path(output_today,"Habitat_batpass_prop_night.tif"), width = 9, height = 6, units = "in", res = 350)

p <- ggplot()
p <- p + xlab("night") + ylab("batpassTOTAL_prop")
p <- p + theme(text = element_text(size = 15))
p <- p + geom_point(data = df3C, 
                    aes(x = night, 
                        y = batpass_prop),
                    size = 1)
p <- p + geom_smooth(data = df3C, 
                     aes(x = night, 
                         y = batpass_prop))

p <- p + facet_grid( ~ Habitat)


print(p)
dev.off()

 

#batpass_prop	versus	temp_max			
# save plot to file 
bmp(file.path(output_today, "Habitat_batpass_prop_temp_max.tif"), width = 9, height = 6, units = "in", res = 350)

p <- ggplot()
p <- p + xlab("temp_max") + ylab("batpassTOTAL_prop")
p <- p + theme(text = element_text(size = 15))
p <- p + geom_point(data = df3C, 
                    aes(x = temp_max, 
                        y = batpass_prop),
                    size = 1)
p <- p + geom_smooth(data = df3C, 
                     aes(x = temp_max, 
                         y = batpass_prop))

p <- p + facet_grid( ~ Habitat)


print(p)
dev.off()

#batpass_prop	versus	temp_mean			
# save plot to file 
bmp(file.path(output_today, "Habitat_batpass_prop_temp_mean.tif"), width = 9, height = 6, units = "in", res = 350)

p <- ggplot()
p <- p + xlab("temp_mean") + ylab("batpassTOTAL_prop")
p <- p + theme(text = element_text(size = 15))
p <- p + geom_point(data = df3C, 
                    aes(x = temp_mean, 
                        y = batpass_prop),
                    size = 1)
p <- p + geom_smooth(data = df3C, 
                     aes(x = temp_mean, 
                         y = batpass_prop))

p <- p + facet_grid( ~ Habitat)


print(p)
dev.off()

#batpass_prop	versus	wind_max			
# save plot to file 
bmp(file.path(output_today, "Habitat_batpass_prop_wind_max.tif"), width = 9, height = 6, units = "in", res = 350)

p <- ggplot()
p <- p + xlab("wind_max") + ylab("batpassTOTAL_prop")
p <- p + theme(text = element_text(size = 15))
p <- p + geom_point(data = df3C, 
                    aes(x = wind_max, 
                        y = batpass_prop),
                    size = 1)
p <- p + geom_smooth(data = df3C, 
                     aes(x = wind_max, 
                         y = batpass_prop))

p <- p + facet_grid( ~ Habitat)


print(p)
dev.off()

#batpass_prop	versus	wind_mean			
# save plot to file 
bmp(file.path(output_today, "Habitat_batpass_prop_wind_mean.tif"), width = 9, height = 6, units = "in", res = 350)

p <- ggplot()
p <- p + xlab("wind_mean") + ylab("batpassTOTAL_prop")
p <- p + theme(text = element_text(size = 15))
p <- p + geom_point(data = df3C, 
                    aes(x = wind_mean, 
                        y = batpass_prop),
                    size = 1)
p <- p + geom_smooth(data = df3C, 
                     aes(x = wind_mean, 
                         y = batpass_prop))

p <- p + facet_grid( ~ Habitat)


print(p)
dev.off()

```


####################################################

## Aggregated to night by guild and behavior 
```{r}

bats_gb_nights_sum <- bats1 %>% group_by(Site, night, guild, behavior) %>% dplyr::summarize(batpass = sum(n()))
summary(bats_gb_nights_sum)

bats_gb_nights_sum1 <- bats_gb_nights_sum
summary(bats_gb_nights_sum1)
summary(bats_gb_nights_sum1$Site)
# Met45 Met95   N02   N04   N08   N09   N10   N11   N14   P02   P04   P08   P09   P10   P11   P14 
#    68    20   140   132   142   167   181   138   158   134   117   114   185   162   174   139 

# write.csv(bats_gb_nights_sum1, "bat pass aggregated by night wout zeronights_behavior and guild.csv")


## Insert zero activity nights ##
# same active.nights objects as above
summary(active.nights)
dat_gb <- bats_gb_nights_sum1
summary(dat_gb)
names(dat_gb)
# "Site"     "night"    "guild"    "behavior" "batpass" 

#---- ADDING MISSING DAYS 

# adjust dataset to include nights when detectors were active but no data was collected. 

head(active.nights)
summary(active.nights$Site)

nights <- unique(active.nights$night) # 91 unique nights of survey 

temp <- expand.grid(Site = unique(dat_gb$Site), 
                  guild = unique(dat_gb$guild), 
                  behavior = unique(dat_gb$behavior),  
                  night = nights) #----expanded data

temp$night <- as.character(temp$night)
temp$night <-  as.Date(temp$night)

active.nights$Site <- as.factor(active.nights$Site)
active.nights$night <-  as.Date(active.nights$night)

str(temp)
str(active.nights)

temp2<-merge(temp, active.nights) # 811 obs of 3 variables 

summary(temp2$Site) # number of monitoring nights per site 
# Met45 Met95   N02   N04   N08   N09   N10   N11   N14   P02   P04   P08   P09   P10   P11   P14 
#   840   840   708   528   720   576   720   720   720   720   720   720   720   720   720   720 

temp3<-merge(temp2, dat_gb, all.x=TRUE) # 811 obs of 6 vars 

summary(temp3$Site)# number of nights per guild per site plot 
# Met45 Met95   N02   N04   N08   N09   N10   N11   N14   P02   P04   P08   P09   P10   P11   P14 
#   840   840   708   528   720   576   720   720   720   720   720   720   720   720   720   720 

## Compare to the original dataset
summary(dat_gb$Site)
# Met45 Met95   N02   N04   N08   N09   N10   N11   N14   P02   P04   P08   P09   P10   P11   P14 
#    68    20   140   132   142   167   181   138   158   134   117   114   185   162   174   139 

summary(temp3) # 9241 zero nights/behavior/guilds introduced 


# reformat factor columns and then replace NA bat passes with zeros 
nbats1_gb <- temp3 
# batpasses converted to zeros but NAs pop up in the factor columns
# fix Habitat and Facility 

nbats2_gb <- nbats1_gb %>% select(-c(active.night)) 
summary(nbats2_gb)  

#Replace NAs with zeros 
nbats2_gb[is.na(nbats2_gb)] <- 0 
summary(nbats2_gb)

sum(nbats2$batpass) 
# 19438 # good! 

nbats3_gb <- left_join(nbats2_gb, bat_sitemap) 
summary(nbats3_gb) # 11412 obs of 8 variables 

# write.csv(nbats3_gb, "nightlyaggregatedMarker2020_totalbats_zeroinserted_behavior and guild.csv")

```

## Aggregate the hourly weather datasets to nightly 
```{r}

head(weather.nights)

nbats4_gb <- left_join(nbats3_gb, weather.nights, by = "night")

summary(nbats4_gb)# There is an index column that can be removed 
nbats4_gb <- nbats4_gb %>% select(-"...1")


```

## Create binary bat pass value (nights with and without activity)
```{r}
bats_gb <- nbats4_gb %>% rename(Batpass = batpass)
bats_gb$batpass01 <- bats_gb$Batpass
names(bats_gb)
 # [1] "Site"        "night"       "batpass_sum" "Habitat"     "Facility"    "avg_wind"    "min_wind"    "max_wind"   
 # [9] "avg_temp"    "min_temp"    "max_temp"    "batpass01"  

bats_gb$batpass01[bats_gb$batpass01 >1] <- 1
summary(bats_gb) # 11412 obs of 15 vars 
head(bats_gb)
tail(bats_gb)

# * Main results 
# write.csv(bats_gb,"nightlyaggregated_zeros_binary_behavior and guild.csv")

# last rounds of housekeeping
bats_gb$wind <- bats_gb$avg_wind
bats_gb$temp <- bats_gb$avg_temp
bats_gb$jnight <- yday(bats_gb$night)

# Aggregation table 
agg_tbl2 <- bats_gb %>% group_by(night, Site, guild, behavior) %>% 
  dplyr::summarise(across(c(Batpass, batpass01, wind, temp), list(sum = sum, mean = mean, min = min, max=max, length = length)))

dim(agg_tbl2)
# 11412    24
head(agg_tbl2)
tail(agg_tbl2)

#Remove variables that will not be needed
names(agg_tbl2)
#  [1] "jnight"           "Site"             "guild"            "behavior"         "Batpass_sum"      "Batpass_mean"    
#  [7] "Batpass_min"      "Batpass_max"      "Batpass_length"   "batpass01_sum"    "batpass01_mean"   "batpass01_min"   
# [13] "batpass01_max"    "batpass01_length" "wind_sum"         "wind_mean"        "wind_min"         "wind_max"        
# [19] "wind_length"      "temp_sum"         "temp_mean"        "temp_min"         "temp_max"         "temp_length"


df2C <- agg_tbl2 %>% select(-c(Batpass_min, Batpass_length, batpass01_mean, batpass01_min, batpass01_max,
                                wind_sum, wind_length, temp_sum , temp_length ))

dim(df2C) 
# 11412    15

head(df2C)
tail(df2C)


#Make a proportion variable

df2C$batpass_prop <- df2C$batpass01_sum/df2C$batpass01_length

head(df2C)
tail(df2C)
dim(df2C)
summary(df2C) # is also only ones or zeros... why apply this to the binary variable instead of the bat pass sum? 
sum(df2C$Batpass_sum)
# 19438 we have the correct number of bat passes so this looks good! 

# write.csv(df2C, "aggregated night behavior guild_summarytable_withbinary_batpass.csv")

# Add back the other site data 

df3C <- left_join(df2C, bat_sitemap)
head(df3C) # 11412 obs of 19 vars 

# * Main results 
# write.csv(df3C, "guild_behavior_summarytable_site.csv")

```


## Exploratory figures 
```{r}
# summary of survey effort
names(df3C)
#  [1] "jnight"           "Site"             "guild"            "behavior"         "Batpass_sum"      "Batpass_mean"     "Batpass_max"     
#  [8] "batpass01_sum"    "batpass01_length" "wind_mean"        "wind_min"         "wind_max"         "temp_mean"        "temp_min"        
# [15] "temp_max"         "batpass_prop"     "Habitat"          "Facility"    
summary(df3C)


bmp(file.path(output_today,"Survey effort per site.tif"), width = 9, height = 6, units = "in", res = 350)

# Overview of nights with active detectors and bat passes detected - Figure S1 a. 
figS1a <- 
  ggplot(data=df3C[df3C$Batpass_sum>0,]) +
  geom_point(data=df3C, aes(x=night,y=guild),shape=19,color="grey") +
  geom_point(aes(x=night,y=guild),shape=19,color="black") +
  facet_wrap(~Site)
print(p)
dev.off()

figS1a

```

## Prepare data for downstream modelling 
```{r}
# Remove Met tower , MRE bats, and social calls. 

df1 <- df3C %>%  
  dplyr::filter(Habitat != "MeteorologicalTower") %>% 
  dplyr::filter(guild != "MRE") %>% 
  dplyr::filter(behavior != "Social") %>% 
  droplevels() 

summary(df1) # 4866 obs of 18 vars

# * Main results 
# write.csv(df1, "guild_behavior_batpass_summarytable_trimmed MRE social and met tower.csv") 

```

## New summary plots 
```{r}
names(df1) 

#  [1] "night"            "Site"             "guild"            "behavior"         "Batpass_sum"      "Batpass_mean"    
#  [7] "Batpass_max"      "batpass01_sum"    "batpass01_length" "wind_mean"        "wind_min"         "wind_max"        
# [13] "temp1_mean"        "temp1_min"         "temp1_max"         "batpass_prop"     "Habitat"          "Facility"   

#batpass_prop	versus	night
# save plot to file 
bmp(file.path(output_today, "Habitat_batpass_prop_night.tif"), width = 9, height = 6, units = "in", res = 350)

p <- ggplot()
p <- p + xlab("night") + ylab("batpassTOTAL_prop")
p <- p + theme(text = element_text(size = 15))
p <- p + geom_point(data = df1, 
                    aes(x = night, 
                        y = batpass_prop),
                    size = 1)
p <- p + geom_smooth(data = df1, 
                     aes(x = night, 
                         y = batpass_prop))

p <- p + facet_grid( behavior ~ Habitat)


print(p)
dev.off()

 

#batpass_prop	versus	temp1_max			
# save plot to file 
bmp(file.path(output_today,"Habitat_batpass_prop_temp1_max.tif"), width = 9, height = 6, units = "in", res = 350)

p <- ggplot()
p <- p + xlab("temp_max") + ylab("batpassTOTAL_prop")
p <- p + theme(text = element_text(size = 15))
p <- p + geom_point(data = df1, 
                    aes(x = temp_max, 
                        y = batpass_prop),
                    size = 1)
p <- p + geom_smooth(data = df1, 
                     aes(x = temp_max, 
                         y = batpass_prop))

p <- p + facet_grid( behavior ~ Habitat)


print(p)
dev.off()

#batpass_prop	versus	temp1_mean			
# save plot to file 
bmp(file.path(output_today,"Habitat_batpass_prop_temp1_mean.tif"), width = 9, height = 6, units = "in", res = 350)

p <- ggplot()
p <- p + xlab("temp_mean") + ylab("batpassTOTAL_prop")
p <- p + theme(text = element_text(size = 15))
p <- p + geom_point(data = df1, 
                    aes(x = temp_mean, 
                        y = batpass_prop),
                    size = 1)
p <- p + geom_smooth(data = df1, 
                     aes(x = temp_mean, 
                         y = batpass_prop))

p <- p + facet_grid(behavior ~ Habitat)


print(p)
dev.off()

#batpass_prop	versus	wind_max			
# save plot to file 
bmp(file.path(output_today,"Habitat_batpass_prop_wind_max.tif"), width = 9, height = 6, units = "in", res = 350)

p <- ggplot()
p <- p + xlab("wind_max") + ylab("batpassTOTAL_prop")
p <- p + theme(text = element_text(size = 15))
p <- p + geom_point(data = df1, 
                    aes(x = wind_max, 
                        y = batpass_prop),
                    size = 1)
p <- p + geom_smooth(data = df1, 
                     aes(x = wind_max, 
                         y = batpass_prop))

p <- p + facet_grid(behavior ~ Habitat)


print(p)
dev.off()

#batpass_prop	versus	wind_mean			
# save plot to file 
bmp(file.path(output_today,"Habitat_batpass_prop_wind_mean.tif"), width = 9, height = 6, units = "in", res = 350)

p <- ggplot()
p <- p + xlab("wind_mean") + ylab("batpassTOTAL_prop")
p <- p + theme(text = element_text(size = 15))
p <- p + geom_point(data = df1, 
                    aes(x = wind_mean, 
                        y = batpass_prop),
                    size = 1)
p <- p + geom_smooth(data = df1, 
                     aes(x = wind_mean, 
                         y = batpass_prop))

p <- p + facet_grid(behavior ~ Habitat)


print(p)
dev.off()
